Revamp deterministic-wasi fixture
- Put it in its own folder so we can rebuild it
- Build it in release mode to shave off ~1s in test
