+++ spec/fixtures/wasi-deterministic/.cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+target = "wasm32-wasip1"
