+++ spec/fixtures/wasi-deterministic/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "wasi-deterministic"
+version = "0.1.0"
+edition = "2021"
+
+[profile.release]
+codegen-units = 1
+opt-level = "s"
+debug = false
+strip = true
+lto = true
+
+[dependencies]
+chrono = "0.4.38"
+rand = "0.8.5"
+serde = "1.0.210"
+serde_json = "1.0.128"
